Решение 2 задачи с дробными числами

+++ 2.cs
@@ -1,9 +1,9 @@
 ﻿// Напишите программу, которая на вход принимает два числа и выдаёт, какое число большее, а какое меньшее
 Console.Write("Введите первое число: ");
-int firstDigit = Convert.ToInt32(Console.ReadLine());
+double firstDigit = Convert.ToDouble(Console.ReadLine());
 
 Console.Write("Введите второе число: ");
-int secondDigit = Convert.ToInt32(Console.ReadLine());
+double secondDigit = Convert.ToDouble(Console.ReadLine());
 if (firstDigit > secondDigit)
 {
     Console.WriteLine("Число " + firstDigit + " больше числа " + secondDigit);
